Refactor: add layer2-sequencer-health endpoint folder structure
