+++ packages/sources/layer2-sequencer-health/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as health from './health'
