+++ packages/sources/layer2-sequencer-health/src/index.ts
@@ -1,5 +1,7 @@
 import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute } from './adapter'
+import { makeExecute, endpointSelector } from './adapter'
 import { makeConfig, NAME } from './config'
+import * as types from './endpoint'
 
-export = { NAME, makeExecute, makeConfig, ...expose(NAME, makeExecute()) }
+const { server } = expose(NAME, makeExecute(), undefined, endpointSelector)
+export { NAME, makeExecute, makeConfig, server, types }
